refac: refactoring all types
