+++ types/CapitalInfo.ts
@@ -1,0 +1,3 @@
+export interface CapitalInfo {
+  latlng: number[];
+}
