+++ types/Car.ts
@@ -1,0 +1,4 @@
+export interface Car {
+  signs: string[];
+  side: string;
+}
