+++ types/CoatOfArms.ts
@@ -1,0 +1,4 @@
+export interface CoatOfArms {
+  png: string;
+  svg: string;
+}
