+++ types/Country.ts
@@ -1,3 +1,17 @@
+import { Currencies } from "./Currencies";
+import { Idd } from "./Idd";
+import { Translation } from "./Translation";
+import { Demonyms } from "./Demonyms";
+import { Maps } from "./Maps";
+import { Gini } from "./Gini";
+import { Car } from "./Car";
+import { CoatOfArms } from "./CoatOfArms";
+import { PostalCode } from "./PostalCode";
+import { Languages }  from "@/types/Languages";
+import { Name }  from "@/types/Name";
+import { CapitalInfo }  from "@/types/CapitalInfo";
+
+
 export interface Country {
   name: Name;
   tld: string[];
@@ -35,82 +49,3 @@ export interface Country {
   capitalInfo: CapitalInfo;
   postalCode: PostalCode;
 }
-
-export interface CapitalInfo {
-  latlng: number[];
-}
-
-export interface Car {
-  signs: string[];
-  side: string;
-}
-
-export interface CoatOfArms {
-  png: string;
-  svg: string;
-}
-
-export interface Currencies {
-  USD: Usd;
-}
-
-export interface Usd {
-  name: string;
-  symbol: string;
-}
-
-export interface Demonyms {
-  eng: Eng;
-  fra: Eng;
-}
-
-export interface Eng {
-  f: string;
-  m: string;
-}
-
-export interface Gini {
-  "2018": number;
-}
-
-export interface Idd {
-  root: string;
-  suffixes: string[];
-}
-
-export interface Languages {
-  eng: string;
-  spa: string;
-  por: string;
-  fra: string
-  nld: string
-  kal: string
-}
-
-export interface Maps {
-  googleMaps: string;
-  openStreetMaps: string;
-}
-
-export interface Name {
-  common: string;
-  official: string;
-  nativeName: NativeName;
-}
-
-export interface NativeName {
-  eng: Translation;
-  zho: Translation;
-}
-
-export interface Translation {
-  official: string;
-  common: string;
-}
-
-export interface PostalCode {
-  format: string;
-  regex: string;
-}
-
-// aqui estão as tipagens dos dados que vamos receber na requisição dos países
