+++ types/Currencies.ts
@@ -1,0 +1,5 @@
+import { Usd } from "./Usd";
+
+export interface Currencies {
+  USD: Usd;
+}
