+++ types/Demonyms.ts
@@ -1,0 +1,6 @@
+import { Eng } from "./Eng";
+
+export interface Demonyms {
+  eng: Eng;
+  fra: Eng;
+}
