+++ types/Eng.ts
@@ -1,0 +1,4 @@
+export interface Eng {
+  f: string;
+  m: string;
+}
