+++ types/Gini.ts
@@ -1,0 +1,3 @@
+export interface Gini {
+  "2018": number;
+}
