+++ types/Idd.ts
@@ -1,0 +1,4 @@
+export interface Idd {
+  root: string;
+  suffixes: string[];
+}
