+++ types/Languages.ts
@@ -1,0 +1,8 @@
+export interface Languages {
+  eng: string;
+  spa: string;
+  por: string;
+  fra: string
+  nld: string
+  kal: string
+}
