+++ types/Maps.ts
@@ -1,0 +1,4 @@
+export interface Maps {
+  googleMaps: string;
+  openStreetMaps: string;
+}
