+++ types/Name.ts
@@ -1,0 +1,7 @@
+import { NativeName } from "./NativeName";
+
+export interface Name {
+  common: string;
+  official: string;
+  nativeName: NativeName;
+}
