+++ types/NativeName.ts
@@ -1,0 +1,6 @@
+import { Translation } from "./Translation";
+
+export interface NativeName {
+  eng: Translation;
+  zho: Translation;
+}
