+++ types/PostalCode.ts
@@ -1,0 +1,4 @@
+export interface PostalCode {
+  format: string;
+  regex: string;
+}
