+++ types/Translation.ts
@@ -1,0 +1,4 @@
+export interface Translation {
+  official: string;
+  common: string;
+}
