+++ types/Usd.ts
@@ -1,0 +1,4 @@
+export interface Usd {
+  name: string;
+  symbol: string;
+}
